v1.0

+++ USER/Device/Inc/bmi088.h
@@ -263,13 +263,13 @@ typedef struct
 	int16_t gy;
 	int16_t gz;
 	
-	float ax_offset;
-	float ay_offset;
-	float az_offset;
+	int32_t ax_offset;
+	int16_t ay_offset;
+	int16_t az_offset;
 
-	float gx_offset;
-	float gy_offset;
-	float gz_offset;
+	int16_t gx_offset;
+	int16_t gy_offset;
+	int16_t gz_offset;
 } mpu_data_t;
 
 enum
